Refactor settings folderstructure. Changes to ui
