+++ src/pages/settings/Faq.tsx
@@ -1,16 +1,23 @@
 import { useState } from "react";
-import SearchAndFilter from "../components/common/SearchAndFilter";
-import { Accordion } from "../components/common/Accordion";
-import TextArea from "../components/form-components/TextArea";
-import CustomButton from "../components/form-components/CustomButton";
-import { QuestionsPropsType } from "../types/AllTypes";
-import { help } from "../mocks/AllMockData";
+import SearchAndFilter from "../../components/common/SearchAndFilter";
+import TextArea from "../../components/form-components/TextArea";
+import CustomButton from "../../components/form-components/CustomButton";
+import { Accordion } from "../../components/common/Accordion";
+import { faqs } from "../../mocks/AllMockData";
+import { QuestionsPropsType } from "../../types/AllTypes";
 
-const Help = () => {
+const Faq = () => {
    const [searchInput, setSearchInput] = useState("");
    const [selectedFilter, setSelectedFilter] = useState("");
    const [activeFilter, setActiveFilter] = useState("All");
-   const filterTitleList = ["All", "Ask a question"];
+   const filterTitleList = [
+      "All",
+      "Services",
+      "Audits",
+      "Tax",
+      "Security",
+      "Technology",
+   ];
 
    const handleSearch = (e: React.ChangeEvent<HTMLInputElement>) => {
       setSearchInput(e.target.value);
@@ -25,16 +32,20 @@ const Help = () => {
       setSelectedFilter(title);
    };
 
-   const filterByQuestions = () => {
+   const filterByQuestions = (question: QuestionsPropsType) => {
       if (selectedFilter === "" || selectedFilter === "All") {
          return true;
       }
-      return;
+      return (
+         question.questionCategory?.toLowerCase() ===
+         selectedFilter.toLowerCase()
+      );
    };
 
-   const filterQuestions = help
-      ? help.filter(
-           (question) => searchQuestions(question) && filterByQuestions()
+   const filterQuestions = faqs
+      ? faqs.filter(
+           (question) =>
+              searchQuestions(question) && filterByQuestions(question)
         )
       : [];
 
@@ -43,7 +54,7 @@ const Help = () => {
    return (
       <div className="">
          <div className="h-[600px] flex flex-col gap-7 mt-4 px-2 py-3 overflow-scroll">
-            <h1 className="text-lg font-semibold">Help and support</h1>
+            <h1 className="text-lg font-semibold">FAQs</h1>
 
             <SearchAndFilter
                handleSearch={handleSearch}
@@ -82,4 +93,4 @@ const Help = () => {
    );
 };
 
-export default Help;
+export default Faq;
