+++ src/pages/settings/IntegratedBanks.tsx
@@ -1,0 +1,9 @@
+const IntegratedBanks = () => {
+   return (
+      <div className="h-[600px] flex flex-col gap-7 mt-4 px-2 py-3 overflow-scroll">
+         <h1 className="text-lg font-semibold">Integrated banks</h1>
+      </div>
+   );
+};
+
+export default IntegratedBanks;
