+++ src/pages/settings/MyDocuments.tsx
@@ -1,0 +1,9 @@
+const MyDocuments = () => {
+   return (
+      <div className="h-[600px] flex flex-col gap-7 mt-4 px-2 py-3 overflow-scroll">
+         <h1 className="text-lg font-semibold">My documents</h1>
+      </div>
+   );
+};
+
+export default MyDocuments;
