+++ src/pages/settings/Privacy.tsx
