+++ src/pages/settings/Terms.tsx
